Reorganized agents_and_tools directory
