+++ code_tests/low_cost_or_live_api_tests/test_smart_searcher.py
@@ -3,7 +3,9 @@ import re
 
 import pytest
 
-from forecasting_tools.agents_and_tools.smart_searcher import SmartSearcher
+from forecasting_tools.agents_and_tools.research.smart_searcher import (
+    SmartSearcher,
+)
 from forecasting_tools.ai_models.general_llm import GeneralLlm
 
 logger = logging.getLogger(__name__)
