+++ forecasting_tools/__init__.py
@@ -12,19 +12,19 @@ from forecasting_tools.agents_and_tools.base_rates.niche_list_researcher import 
 from forecasting_tools.agents_and_tools.base_rates.niche_list_researcher import (
     NicheListResearcher as NicheListResearcher,
 )
-from forecasting_tools.agents_and_tools.key_factors_researcher import (
-    KeyFactorsResearcher as KeyFactorsResearcher,
-)
-from forecasting_tools.agents_and_tools.key_factors_researcher import (
-    ScoredKeyFactor as ScoredKeyFactor,
-)
 from forecasting_tools.agents_and_tools.question_generators.question_generator import (
     QuestionGenerator as QuestionGenerator,
 )
 from forecasting_tools.agents_and_tools.question_generators.topic_generator import (
     TopicGenerator as TopicGenerator,
 )
-from forecasting_tools.agents_and_tools.smart_searcher import (
+from forecasting_tools.agents_and_tools.research.key_factors_researcher import (
+    KeyFactorsResearcher as KeyFactorsResearcher,
+)
+from forecasting_tools.agents_and_tools.research.key_factors_researcher import (
+    ScoredKeyFactor as ScoredKeyFactor,
+)
+from forecasting_tools.agents_and_tools.research.smart_searcher import (
     SmartSearcher as SmartSearcher,
 )
 from forecasting_tools.ai_models.ai_utils.ai_misc import (
