+++ forecasting_tools/agents_and_tools/base_rates/deduplicator.py
@@ -11,7 +11,9 @@ from sklearn.metrics.pairwise import cosine_similarity
 from forecasting_tools.agents_and_tools.deprecated.configured_llms import (
     BasicLlm,
 )
-from forecasting_tools.agents_and_tools.smart_searcher import SmartSearcher
+from forecasting_tools.agents_and_tools.research.smart_searcher import (
+    SmartSearcher,
+)
 from forecasting_tools.ai_models.ai_utils.ai_misc import clean_indents
 from forecasting_tools.util.misc import raise_for_status_with_additional_info
