+++ forecasting_tools/agents_and_tools/base_rates/niche_list_researcher.py
@@ -12,7 +12,9 @@ from forecasting_tools.agents_and_tools.base_rates.deduplicator import (
 from forecasting_tools.agents_and_tools.deprecated.configured_llms import (
     BasicLlm,
 )
-from forecasting_tools.agents_and_tools.smart_searcher import SmartSearcher
+from forecasting_tools.agents_and_tools.research.smart_searcher import (
+    SmartSearcher,
+)
 from forecasting_tools.ai_models.ai_utils.ai_misc import clean_indents
 from forecasting_tools.util import async_batching
 from forecasting_tools.util.misc import (
