+++ forecasting_tools/agents_and_tools/deprecated/general_researcher.py
@@ -5,7 +5,9 @@ from forecasting_tools.agents_and_tools.deprecated.question_responder import (
 )
 
 logger = logging.getLogger(__name__)
-from forecasting_tools.agents_and_tools.smart_searcher import SmartSearcher
+from forecasting_tools.agents_and_tools.research.smart_searcher import (
+    SmartSearcher,
+)
 from forecasting_tools.ai_models.ai_utils.ai_misc import clean_indents
