+++ forecasting_tools/agents_and_tools/misc_tools.py
@@ -3,7 +3,9 @@ import asyncio
 from forecasting_tools.agents_and_tools.question_generators.simple_question import (
     SimpleQuestion,
 )
-from forecasting_tools.agents_and_tools.smart_searcher import SmartSearcher
+from forecasting_tools.agents_and_tools.research.smart_searcher import (
+    SmartSearcher,
+)
 from forecasting_tools.ai_models.agent_wrappers import AgentTool, agent_tool
 from forecasting_tools.ai_models.ai_utils.ai_misc import clean_indents
 from forecasting_tools.ai_models.general_llm import GeneralLlm
