+++ forecasting_tools/agents_and_tools/research/find_a_dataset.py
@@ -3,9 +3,11 @@ import logging
 
 from pydantic import BaseModel
 
-from forecasting_tools.agents_and_tools.computer_use import ComputerUse
-from forecasting_tools.agents_and_tools.data_analyzer import DataAnalyzer
 from forecasting_tools.agents_and_tools.misc_tools import perplexity_pro_search
+from forecasting_tools.agents_and_tools.other.data_analyzer import DataAnalyzer
+from forecasting_tools.agents_and_tools.research.computer_use import (
+    ComputerUse,
+)
 from forecasting_tools.ai_models.agent_wrappers import (
     AgentRunner,
     AgentSdkLlm,
