+++ forecasting_tools/agents_and_tools/research/key_factors_researcher.py
@@ -16,7 +16,9 @@ from forecasting_tools.agents_and_tools.deprecated.configured_llms import (
 from forecasting_tools.agents_and_tools.deprecated.research_coordinator import (
     ResearchCoordinator,
 )
-from forecasting_tools.agents_and_tools.smart_searcher import SmartSearcher
+from forecasting_tools.agents_and_tools.research.smart_searcher import (
+    SmartSearcher,
+)
 from forecasting_tools.ai_models.ai_utils.ai_misc import clean_indents
 from forecasting_tools.helpers.metaculus_api import MetaculusQuestion
 from forecasting_tools.util import async_batching
