+++ scripts/run_key_factors.py
@@ -5,7 +5,7 @@ from datetime import datetime
 from forecasting_tools.agents_and_tools.deprecated.general_researcher import (
     GeneralResearcher,
 )
-from forecasting_tools.agents_and_tools.key_factors_researcher import (
+from forecasting_tools.agents_and_tools.research.key_factors_researcher import (
     KeyFactorsResearcher,
     ScoredKeyFactor,
 )
